example/falcon_auth_test/requirements.txt
falcon
gunicorn
python-dotenv
ikon-auth>=0.1.5